+++ CHANGELOG.rst
@@ -44,6 +44,7 @@ Merged branches and closed issues
 * Qblox ICCs - Compensated integration time for Qblox QRM IC component (!199).
 * Qblox ICCs - Added error handling for error flags given by `get_sequencer_state` (!215)
 * Visualization - Allow user defined axis for plotting circuit diagram (!206)
+* Visualization - Adds schedule plotting using matplotlib and a WindowOperation to help visualize pulse diagrams (!225, !232)
 * Added method `sample_schedule` to sample a `Schedule` (!212)
 
 0.4.0 InstrumentCoordinator and improvements to backends (2021-08-06)